backend/requirements.txt
langchain
python-dotenv
langchain-community
langchain-experimental
fastapi
python-socketio
langchain_google_genai
uvicorn
chromadb
google-generativeai
colorlog
jinja2
pydantic
motor
transformers
torch
sentence_transformers
python-multipart
pymongo
python-jose
Werkzeug
PyJWT
passlib
python-decouple